Remove parallelism and reduce RAM usage in plain_word_freqs.py
Load the files in the program directly rather than using separate
processes.  This is unnecessary: for small corpus the gain is not
sufficient to compensate the cost of process creation and IPC.  For
large corpus the parallelism causes unwanted disk movement and hurts
performance.

Once the counts for a file is computed, merge it to the running counts
and discard the file counts.  This is very important for large corpus,
otherwise the dicts for the large number of files easily fill all the
available memory and cause the machine to thrash.

This also makes the display of tqdm much more smooth.

Signed-off-by: Isaac To <isaac.to@gmail.com>